app/pyproject.toml
[tool.poetry]
name = "bellissimo"
version = "0.1.0"
description = "Meal planner for slackers"
authors = ["Cabeda"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.2"
fastapi = "^0.63.0"
uvicorn = "^0.11.8"
sqlalchemy = "^1.3.18"
psycopg2-binary = "^2.8.5"
bcrypt = "^3.2.0"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
black = {version = "^19.10b0", allow-prereleases = true}
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
